Currently stuck until I can learn OCaml funamentals
This is as far as I got before my lack of knowledge really put a
halt on the project. In trying to find a good way to iterate over
the characters of a string I continually get hung up on syntax
errors, and having incorrect types for the various type signatures
I need.